der bliver ikke tilføjet til kurven hvis antal er over det i lageret. den refresher bare siden(ikke brugervenligt).

+++ Projekt Mappe/Website/Gui/Controllers/DrinkController.cs
@@ -10,6 +10,7 @@ using Gui.WalletServiceRef;
 using Gui.FavoritServiceRef;
 using Gui.UserServiceRef;
 using System.Dynamic;
+using Gui.StorageServiceRef;
 
 namespace Gui.Controllers
 {
@@ -21,6 +22,8 @@ namespace Gui.Controllers
         private OrderServiceClient orderClient = new OrderServiceClient();
         private OrderController oCtr = new OrderController();
         private WalletServiceClient walletClient = new WalletServiceClient();
+        private StorageServiceClient sClient = new StorageServiceClient();
+        
 
         // GET: Drink
         public ActionResult Index()
@@ -63,33 +66,35 @@ namespace Gui.Controllers
             Gui.OrderLineServiceRef.OrderLine orderline = new Gui.OrderLineServiceRef.OrderLine();
             orderline.Amount = amount;
 
-            if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
+            if (sClient.getStorageByDrinkAndStorage(drinkId, cusID).Amount >= amount)
             {
-                if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
                 {
-                    orderline.Drink = lc.GetDrink(drinkId);
-                    orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
-                    lc.CreateOrderLine(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
+                    if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                    {
+                        orderline.Drink = lc.GetDrink(drinkId);
+                        orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
+                        lc.CreateOrderLine(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
 
-                    
-                    decimal hej = orderline.Drink.Price * orderline.Amount;
-                    hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
-                    orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
-                    return RedirectToAction("Details", "Customer", new { id = cusID });
+
+                        decimal hej = orderline.Drink.Price * orderline.Amount;
+                        hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
+                        orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
+                        return RedirectToAction("Details", "Customer", new { id = cusID });
+                    }
+                    else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                    {
+                        orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
+                        oCtr.CreateOrder(cusID);
+                        return Create(amount, drinkId, cusID, UserName);
+                    }
                 }
-                else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                else
                 {
-                    orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
                     oCtr.CreateOrder(cusID);
                     return Create(amount, drinkId, cusID, UserName);
                 }
             }
-            else
-            {
-                oCtr.CreateOrder(cusID);
-                return Create(amount, drinkId, cusID, UserName);
-            }
-
             return RedirectToAction("Details", "Customer", new { id = cusID });
         }
         public ActionResult CreateHelflask()
@@ -104,31 +109,35 @@ namespace Gui.Controllers
             Gui.OrderLineServiceRef.OrderLine orderline = new Gui.OrderLineServiceRef.OrderLine();
             orderline.Amount = amount;
 
-            if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
+            if (sClient.getHelflaskStorageByHelflaskAndStorage(drinkId, cusID).Amount >= amount)
             {
-                if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
                 {
-                    orderline.Drink = lc.GetHelflask(drinkId);
-                    orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
-                    lc.CreateOrderLineHelflask(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
+                    if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                    {
+                        orderline.Drink = lc.GetHelflask(drinkId);
+                        orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
+                        lc.CreateOrderLineHelflask(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
 
 
-                    decimal hej = orderline.Drink.Price * orderline.Amount;
-                    hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
-                    orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
-                    return RedirectToAction("Details", "Customer", new { id = cusID });
+                        decimal hej = orderline.Drink.Price * orderline.Amount;
+                        hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
+                        orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
+                        return RedirectToAction("Details", "Customer", new { id = cusID });
+                    }
+                    else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                    {
+                        orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
+                        oCtr.CreateOrder(cusID);
+                        return CreateHelflask(amount, drinkId, cusID, UserName);
+                    }
                 }
-                else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                else
                 {
-                    orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
                     oCtr.CreateOrder(cusID);
                     return CreateHelflask(amount, drinkId, cusID, UserName);
                 }
-            }
-            else
-            {
-                oCtr.CreateOrder(cusID);
-                return CreateHelflask(amount, drinkId, cusID, UserName);
+
             }
 
             return RedirectToAction("Details", "Customer", new { id = cusID });
@@ -146,33 +155,35 @@ namespace Gui.Controllers
             Gui.OrderLineServiceRef.OrderLine orderline = new Gui.OrderLineServiceRef.OrderLine();
             orderline.Amount = amount;
 
-            if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
+            if (sClient.getAlchoholStorageByDrinkAndStorage(drinkId, cusID).Amount >= amount)
             {
-                if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                if (orderClient.GetOrderByStatus("Incomplete") != null && orderClient.GetUser(UserName).ID == orderClient.GetOrderByStatus("Incomplete").User.ID)
                 {
-                    orderline.Drink = lc.GetAlchohol(drinkId);
-                    orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
-                    lc.CreateOrderLineAlchohol(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
+                    if (orderClient.GetOrderByStatus("Incomplete").Customer.ID == cusID)
+                    {
+                        orderline.Drink = lc.GetAlchohol(drinkId);
+                        orderline.TotalPrice = orderline.Drink.Price * orderline.Amount;
+                        lc.CreateOrderLineAlchohol(orderline, orderClient.GetOrderByStatus("Incomplete").ID);
 
 
-                    decimal hej = orderline.Drink.Price * orderline.Amount;
-                    hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
-                    orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
-                    return RedirectToAction("Details", "Customer", new { id = cusID });
+                        decimal hej = orderline.Drink.Price * orderline.Amount;
+                        hej = orderClient.GetOrderByStatus("Incomplete").TotalPrice + hej;
+                        orderClient.UpdatePrice(orderClient.GetOrderByStatus("Incomplete"), hej);
+                        return RedirectToAction("Details", "Customer", new { id = cusID });
+                    }
+                    else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                    {
+                        orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
+                        oCtr.CreateOrder(cusID);
+                        return CreateAlchohol(amount, drinkId, cusID, UserName);
+                    }
                 }
-                else if (orderClient.GetOrderByStatus("Incomplete").Customer.ID != cusID)
+                else
                 {
-                    orderClient.DeleteOrderByID(orderClient.GetOrderByStatus("Incomplete").ID);
                     oCtr.CreateOrder(cusID);
                     return CreateAlchohol(amount, drinkId, cusID, UserName);
                 }
             }
-            else
-            {
-                oCtr.CreateOrder(cusID);
-                return CreateAlchohol(amount, drinkId, cusID, UserName);
-            }
-
             return RedirectToAction("Details", "Customer", new { id = cusID });
         }
